backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Cors
python-dotenv
gunicorn
psycopg2-binary